Refactoring BlueprintPage Components
